pyproject.toml
[project]
name = "react-newsletter"
version = "0.1.0"
description = "AI News Newsletter - Multi-agent system for collecting, summarizing, and presenting AI news"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "svlearn-bootcamp>=0.1.7",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "cachetools>=5.3.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "openai>=1.3.0",
    "crewai>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "serpapi>=0.1.5",
    "pillow>=10.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
]
jupyter = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "S101", "PLR0913", "PLR0912", "PLR0915"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/static",
    "/templates",
    # ... other includes like README, pyproject.toml, etc.
]

[tool.hatch.build.targets.wheel]
packages = ["src/ReactNewslettr"]
